Added chat component system

+++ Core/Lib/Scenes/Ingame/IngameRenderContext.cs
@@ -7,13 +7,13 @@ namespace Core.Scenes.Ingame;
 public class IngameRenderContext: IRenderContext
 {
     public Vector2 BaseScreenSize { get; }
-    public float ChatWidth { get; }
+    public int ChatWidth { get; }
     public Color BackgroundColor { get; }
     public RectangleF WorldCulling { get; }
 
     public IngameRenderContext(
         Vector2 baseScreenSize,
-        float chatWidth,
+        int chatWidth,
         Color backgroundColor, 
         RectangleF worldCulling
     )
